riscv64: Support IMSIC
Send IPI via imsic.
