+++ src/arch/riscv64/kernel/interrupts/aplic.rs
@@ -1,0 +1,600 @@
+//! Advanced Platform-Level Interrupt Controller (APLIC) driver for RISC-V.
+//!
+//! [RISC-V Advanced Interrupt Architecture]: https://github.com/riscv/riscv-aia/releases/download/20250312/riscv-interrupts-20250312.pdf
+
+use core::mem::offset_of;
+use core::num::NonZeroU16;
+use core::ptr::NonNull;
+
+use align_address::Align;
+use bitfield_struct::bitfield;
+use free_list::PageLayout;
+use memory_addresses::{PhysAddr, VirtAddr};
+use volatile::access::{NoAccess, ReadOnly, ReadWrite, WriteOnly};
+use volatile::{VolatileFieldAccess, VolatileRef};
+
+use crate::arch::kernel::devicetree::InterruptType as DeviceTreeInterruptType;
+use crate::arch::kernel::interrupts::{EXTERNAL_INTERRUPT_CONTROLLER, ExternalInterruptController};
+use crate::arch::mm::paging::{self, BasePageSize, PageSize, PageTableEntryFlags};
+use crate::arch::riscv64::kernel::core_local::{core_id, msi_controller};
+use crate::mm::{PageAlloc, PageRangeAllocator};
+
+#[bitfield(u32)]
+struct DomainConfig {
+	// BE - Big Endian (0 = little endian, 1 = big endian)
+	#[bits(1)]
+	big_endian: bool,
+
+	#[bits(1)]
+	__: bool,
+
+	// DM - Domain Mode (0 = direct delivery, 1 = msi delivery)
+	#[bits(1)]
+	msi_delivery: bool,
+
+	#[bits(5)]
+	__: u8,
+
+	// IE - Interrupt Enable (0 = disabled, 1 = enabled)
+	#[bits(1)]
+	interrupt_enable: bool,
+
+	#[bits(23)]
+	__: u32,
+}
+
+#[bitfield(u32)]
+struct SourceConfigRegister {
+	#[bits(10)]
+	__payload: u16,
+
+	#[bits(1)]
+	delegated: u8,
+
+	#[bits(21)]
+	__: u32,
+}
+
+#[derive(Debug)]
+#[repr(u8)]
+pub(crate) enum SourceMode {
+	// Inactive in this domain (and not delegated)
+	Inactive = 0,
+	// Active, detached from the source wire
+	Detached = 1,
+	// Edge-sensitive; interrupt asserted on rising edge
+	Edge1 = 4,
+	// Edge-sensitive; interrupt asserted on falling edge
+	Edge0 = 5,
+	// Level-sensitive; interrupt asserted when high
+	Level1 = 6,
+	// Level-sensitive; interrupt asserted when low
+	Level0 = 7,
+}
+impl SourceMode {
+	const fn into_bits(self) -> u8 {
+		self as _
+	}
+
+	const fn from_bits(bits: u8) -> Self {
+		match bits {
+			0 => Self::Inactive,
+			1 => Self::Detached,
+			4 => Self::Edge1,
+			5 => Self::Edge0,
+			6 => Self::Level1,
+			7 => Self::Level0,
+			_ => panic!("invalid SourceMode bits"),
+		}
+	}
+}
+impl From<DeviceTreeInterruptType> for SourceMode {
+	fn from(value: DeviceTreeInterruptType) -> Self {
+		match value {
+			DeviceTreeInterruptType::None => Self::Inactive,
+			DeviceTreeInterruptType::LowToHighEdge => Self::Edge1,
+			DeviceTreeInterruptType::ActiveLowLevel => Self::Level0,
+			DeviceTreeInterruptType::ActiveHighLevel => Self::Level1,
+			DeviceTreeInterruptType::HighToLowEdge => Self::Edge0,
+		}
+	}
+}
+
+#[bitfield(u32)]
+struct SourceConfigSourceMode {
+	#[bits(3)]
+	mode: SourceMode,
+
+	#[bits(7)]
+	__: u8,
+
+	#[bits(1)]
+	delegated: u8,
+
+	#[bits(21)]
+	__: u32,
+}
+
+#[bitfield(u32)]
+struct SourceConfigDelegated {
+	#[bits(10)]
+	child_index: u16,
+
+	#[bits(1)]
+	delegated: u8,
+
+	#[bits(21)]
+	__: u32,
+}
+
+#[bitfield(u32)]
+struct TargetDirectDelivery {
+	// Priority number for interrupt source
+	//
+	// Lower values indicate higher priority. The maximum priority is 0, and the minimum
+	// priority is 255.
+	#[bits(8)]
+	priority: u8,
+
+	#[bits(10)]
+	__: u16,
+
+	// Hart to which the interrupt for this source will be delivered
+	#[bits(14)]
+	hart_index: u16,
+}
+
+#[bitfield(u32)]
+struct TargetMsiDelivery {
+	// External interrupt identity
+	#[bits(11)]
+	eiid: u16,
+
+	#[bits(1)]
+	__: u8,
+
+	// Number of the target hart’s guest interrupt file to which MSIs will be sent.
+	// Only relevant if domain’s harts implement hypervisor extension.
+	#[bits(6)]
+	__guest_index: u8,
+
+	// Hart to which the interrupt for this source will be forwarded
+	#[bits(14)]
+	hart_index: u16,
+}
+
+#[bitfield(u32)]
+struct TargetRegister {
+	#[bits(18)]
+	__payload: u32,
+
+	// Target hart index
+	#[bits(14)]
+	hart_index: u16,
+}
+
+#[repr(C)]
+#[derive(VolatileFieldAccess)]
+struct AplicControlRegion {
+	// Domain configuration
+	// Bit 8: IE - Interrupt Enable
+	// Bit 2: DM - Domain Mode (0 = direct delivery, 1 = msi delivery)
+	// Bit 0: BE - Big Endian
+	#[access(ReadWrite)]
+	domaincfg: DomainConfig,
+	#[access(ReadWrite)]
+	sourcecfg: [SourceConfigRegister; 1023],
+	#[access(NoAccess)]
+	_reserved_sourcecfg: [u32; (0x1bc0 - 0x1000) / size_of::<SourceConfigRegister>()],
+
+	// MSI address configuration registers (m-mode only)
+	#[access(NoAccess)]
+	_mmsiaddrcfg: u32,
+	#[access(NoAccess)]
+	_mmsiaddrcfgh: u32,
+	#[access(NoAccess)]
+	_smsiaddrcfg: u32,
+	#[access(NoAccess)]
+	_smsiaddrcfgh: u32,
+	#[access(NoAccess)]
+	_reserved_msiaddrcfg: [u32; (0x1c00 - 0x1bd0) / size_of::<u32>()],
+
+	// Set interrupt pending bits
+	#[access(ReadWrite)]
+	_setip: [u32; 32],
+	#[access(NoAccess)]
+	_reserved_setip: [u32; (0x1cdc - 0x1c80) / size_of::<u32>()],
+	// Set interrupt pending bit number
+	#[access(WriteOnly)]
+	_setipnum: u32,
+	#[access(NoAccess)]
+	_reserved_setipnum: [u32; (0x1d00 - 0x1ce0) / size_of::<u32>()],
+
+	// Clear interrupt pending bits
+	#[access(ReadWrite)]
+	_in_clrip: [u32; 32],
+	#[access(NoAccess)]
+	_reserved_in_clrip: [u32; (0x1ddc - 0x1d80) / size_of::<u32>()],
+	// Clear interrupt pending bit number
+	#[access(WriteOnly)]
+	_clripnum: u32,
+	#[access(NoAccess)]
+	_reserved_clripnum: [u32; (0x1e00 - 0x1de0) / size_of::<u32>()],
+
+	// Set interrupt enable bits
+	#[access(ReadWrite)]
+	_setie: [u32; 32],
+	#[access(NoAccess)]
+	_reserved_setie: [u32; (0x1edc - 0x1e80) / size_of::<u32>()],
+	// Set interrupt enable bit number
+	#[access(WriteOnly)]
+	setienum: u32,
+	#[access(NoAccess)]
+	_reserved_setienum: [u32; (0x1f00 - 0x1ee0) / size_of::<u32>()],
+
+	// Clear interrupt enable bits
+	#[access(WriteOnly)]
+	_clrie: [u32; 32],
+	#[access(NoAccess)]
+	_reserved_clrie: [u32; (0x1fdc - 0x1f80) / size_of::<u32>()],
+	// Clear interrupt enable bit number
+	#[access(WriteOnly)]
+	clrienum: u32,
+	#[access(NoAccess)]
+	_reserved_clrienum: [u32; (0x2000 - 0x1fe0) / size_of::<u32>()],
+
+	// Set interrupt pending bit by number (little-endian)
+	#[access(WriteOnly)]
+	_setipnum_le: u32,
+	// Set interrupt pending bit by number (big-endian)
+	#[access(WriteOnly)]
+	_setipnum_be: u32,
+	#[access(NoAccess)]
+	_reserved_setipnum_be: [u32; (0x3000 - 0x2008) / size_of::<u32>()],
+
+	// Generate MSI
+	#[access(ReadWrite)]
+	_genmsi: u32,
+
+	// Interrupt Targets
+	target: [TargetRegister; 1023],
+}
+const _: () = assert!(offset_of!(AplicControlRegion, _mmsiaddrcfg) == 0x1bc0);
+const _: () = assert!(offset_of!(AplicControlRegion, _smsiaddrcfg) == 0x1bc8);
+const _: () = assert!(offset_of!(AplicControlRegion, _setip) == 0x1c00);
+const _: () = assert!(offset_of!(AplicControlRegion, _setipnum) == 0x1cdc);
+const _: () = assert!(offset_of!(AplicControlRegion, _in_clrip) == 0x1d00);
+const _: () = assert!(offset_of!(AplicControlRegion, _clripnum) == 0x1ddc);
+const _: () = assert!(offset_of!(AplicControlRegion, _setie) == 0x1e00);
+const _: () = assert!(offset_of!(AplicControlRegion, setienum) == 0x1edc);
+const _: () = assert!(offset_of!(AplicControlRegion, _clrie) == 0x1f00);
+const _: () = assert!(offset_of!(AplicControlRegion, clrienum) == 0x1fdc);
+const _: () = assert!(offset_of!(AplicControlRegion, _setipnum_le) == 0x2000);
+const _: () = assert!(offset_of!(AplicControlRegion, _genmsi) == 0x3000);
+const _: () = assert!(size_of::<AplicControlRegion>() == 0x4000);
+
+#[bitfield(u32)]
+struct TopInterrupt {
+	#[bits(8)]
+	priority: u8,
+
+	#[bits(8)]
+	__: u8,
+
+	#[bits(10)]
+	identity: u16,
+
+	#[bits(6)]
+	__: u8,
+}
+
+#[repr(C)]
+#[derive(VolatileFieldAccess)]
+struct InterruptDeliveryControl {
+	// Interrupt delivery enable (0 = disabled, 1 = enabled)
+	#[access(ReadWrite)]
+	idelivery: u32,
+
+	// Interrupt force (0 = no effect, 1 = force interrupt)
+	// For testing only.
+	#[access(ReadWrite)]
+	iforce: u32,
+
+	// Interrupt enable threshold (0 = all interrupts enabled, n = interrupts with priority > n enabled)
+	#[access(ReadWrite)]
+	ithreshold: u32,
+
+	#[access(NoAccess)]
+	_reserved: [u32; 3],
+
+	// Top interrupt
+	#[access(ReadOnly)]
+	topi: TopInterrupt,
+
+	// Claim top interrupt
+	#[access(ReadOnly)]
+	claimi: TopInterrupt,
+}
+const _: () = assert!(offset_of!(InterruptDeliveryControl, topi) == 0x18);
+const _: () = assert!(offset_of!(InterruptDeliveryControl, claimi) == 0x1c);
+const _: () = assert!(size_of::<InterruptDeliveryControl>() == 0x20);
+
+const APLIC_DIRECT_DELIVERY_MODE_MAX_HARTS: u16 = 512;
+
+// In direct delivery mode the number of harts is limited to 512.
+type InterruptDeliveryControlArray =
+	[InterruptDeliveryControl; APLIC_DIRECT_DELIVERY_MODE_MAX_HARTS as usize];
+
+pub(crate) struct Aplic {
+	control_region: VolatileRef<'static, AplicControlRegion>,
+	interrupt_delivery_control: VolatileRef<'static, InterruptDeliveryControlArray>,
+	ipriolen: u8,
+}
+
+impl Aplic {
+	fn new(
+		control_region: VolatileRef<'static, AplicControlRegion>,
+		interrupt_delivery_control: VolatileRef<'static, InterruptDeliveryControlArray>,
+	) -> Self {
+		Self {
+			control_region,
+			interrupt_delivery_control,
+			ipriolen: 0,
+		}
+	}
+
+	fn init(&mut self, msi_delivery: bool) {
+		let aplic_ptr = self.control_region.as_mut_ptr();
+		let mut domaincfg = aplic_ptr.domaincfg().read();
+		if domaincfg.big_endian() {
+			domaincfg.set_big_endian(false);
+			aplic_ptr.domaincfg().write(domaincfg);
+			assert!(
+				!aplic_ptr.domaincfg().read().big_endian(),
+				"Only little-endian is supported"
+			);
+		}
+
+		let aplic_ptr = self.control_region.as_mut_ptr();
+		let mut domaincfg = aplic_ptr.domaincfg().read();
+
+		if domaincfg.msi_delivery() != msi_delivery {
+			domaincfg.set_msi_delivery(msi_delivery);
+			aplic_ptr.domaincfg().write(domaincfg);
+			assert!(
+				aplic_ptr.domaincfg().read().msi_delivery() == msi_delivery,
+				"APLIC does not support the desired delivery mode."
+			);
+		}
+
+		let aplic_ptr = self.control_region.as_mut_ptr();
+		aplic_ptr.domaincfg().update(|mut cfg| {
+			cfg.set_interrupt_enable(true);
+			cfg
+		});
+
+		if !msi_delivery {
+			self.ipriolen = self.probe_ipriolen();
+			let hart_idc = unsafe {
+				self.interrupt_delivery_control
+					.as_mut_ptr()
+					.map(|control| control.cast().offset(Aplic::get_hart_index() as isize))
+			};
+			hart_idc.idelivery().write(1);
+		}
+	}
+
+	/// Determines the number of implemented interrupt priority bits (IPRIOLEN)
+	/// by probing the first source's target WARL register.
+	fn probe_ipriolen(&mut self) -> u8 {
+		let saved_source_mode = self.get_interrupt_source_mode(1);
+		self.set_interrupt_source_mode(1, SourceMode::Detached);
+		let target = unsafe {
+			self.control_region
+				.as_mut_ptr()
+				.target()
+				.map(|slice| slice.cast::<TargetRegister>())
+		};
+		let saved_target_reg = target.read();
+		target.write(TargetRegister::from(
+			TargetDirectDelivery::new()
+				.with_hart_index(Aplic::get_hart_index())
+				.with_priority(0xff)
+				.into_bits(),
+		));
+		let readback = TargetDirectDelivery::from(target.read().into_bits()).priority();
+		target.write(saved_target_reg);
+		self.set_interrupt_source_mode(1, saved_source_mode);
+
+		let ipriolen = (u8::BITS - readback.leading_zeros()) as u8;
+		assert!(
+			(1..=8).contains(&ipriolen),
+			"APLIC IPRIOLEN must be between 1 and 8, but probed {ipriolen}"
+		);
+		ipriolen
+	}
+
+	fn get_hart_index() -> u16 {
+		// The core identifier and the hart index of an core in an interrupt domain are not necessarily the same.
+		// The hart index can be extracted from the devicetree as following.
+		// 1. Find all core nodes cpu@X
+		// 2. For each core find core local interrupter node (<compatible> = "riscv,cpu-intc") and get its phandle
+		// 3. Find aplic node (<compatible> = "riscv,aplic", <status> != "disabled") for active interrupt domain
+		// 4. Property <interrupts-extended> is a list of tuples with the following format: <phandle cpu-intc> <interrupt-specifier> ...
+		//    For supervisor external interrupts interrupt-specifier = 0x9
+		//    For machine external interrupts interrupt-specifier = 0xb
+		// 5. The index of the tuple in the list is the hart index of the core in the interrupt domain.
+		//
+		// The core identifier is identical to the interrupt domain hart index if
+		// - the core identifier are continuous and start with 0 and
+		// - the interrupt-extended property of the aplic node is ordered by core identifier.
+		// On QEMU virt machine these assumptions hold true.
+
+		let core_id: u16 = core_id().try_into().unwrap();
+		assert!(
+			core_id < APLIC_DIRECT_DELIVERY_MODE_MAX_HARTS,
+			"APLIC direct delivery mode supports only {APLIC_DIRECT_DELIVERY_MODE_MAX_HARTS} harts, but core_id is {core_id}"
+		);
+		core_id
+	}
+
+	pub fn set_enable_bit(&mut self, irq_number: u16, value: bool) {
+		if value {
+			self.control_region
+				.as_mut_ptr()
+				.setienum()
+				.write(u32::from(irq_number));
+		} else {
+			self.control_region
+				.as_mut_ptr()
+				.clrienum()
+				.write(u32::from(irq_number));
+		}
+
+		if let Some(msi_controller) = msi_controller() {
+			let target: u32 = unsafe {
+				self.control_region
+					.as_mut_ptr()
+					.target()
+					.map(|slice| {
+						slice
+							.cast()
+							.offset(isize::try_from(irq_number).unwrap() - 1)
+					})
+					.read()
+			};
+			let eiid = TargetMsiDelivery::from(target).eiid();
+			msi_controller.set_interrupt_enable(NonZeroU16::new(eiid).unwrap(), value);
+		}
+	}
+
+	pub fn set_interrupt_source_mode(&mut self, irq_number: u16, mode: SourceMode) {
+		let sourcecfg = unsafe {
+			self.control_region.as_mut_ptr().sourcecfg().map(|slice| {
+				slice
+					.cast()
+					.offset(isize::try_from(irq_number).unwrap() - 1)
+			})
+		};
+		sourcecfg.write(SourceConfigRegister::from(
+			SourceConfigSourceMode::new().with_mode(mode).into_bits(),
+		));
+	}
+
+	pub fn get_interrupt_source_mode(&mut self, irq_number: u16) -> SourceMode {
+		let sourcecfg = unsafe {
+			self.control_region.as_mut_ptr().sourcecfg().map(|slice| {
+				slice
+					.cast()
+					.offset(isize::try_from(irq_number).unwrap() - 1)
+			})
+		};
+		let sourcecfg_value: u32 = sourcecfg.read();
+		SourceConfigSourceMode::from(sourcecfg_value).mode()
+	}
+
+	pub fn set_interrupt_priority(&mut self, irq_number: u16, priority: u8) {
+		let new_value = if msi_controller().is_some() {
+			warn!("APLIC interrupt priority is not supported in MSI delivery mode");
+			TargetMsiDelivery::new()
+				.with_hart_index(Aplic::get_hart_index())
+				.with_eiid(irq_number)
+				.into_bits()
+		} else {
+			let max_priority = u8::MAX >> (8 - self.ipriolen);
+			TargetDirectDelivery::new()
+				.with_hart_index(Aplic::get_hart_index())
+				.with_priority(priority.min(max_priority))
+				.into_bits()
+		};
+
+		let target = unsafe {
+			self.control_region.as_mut_ptr().target().map(|slice| {
+				slice
+					.cast()
+					.offset(isize::try_from(irq_number).unwrap() - 1)
+			})
+		};
+		target.write(TargetRegister::from(new_value));
+	}
+
+	pub fn set_priority_threshold(&mut self, threshold: u8) {
+		if let Some(msi_controller) = msi_controller() {
+			msi_controller.set_interrupt_priority_threshold(threshold);
+		} else {
+			let hart_idc = unsafe {
+				self.interrupt_delivery_control
+					.as_mut_ptr()
+					.map(|control| control.cast().offset(Aplic::get_hart_index() as isize))
+			};
+			hart_idc.ithreshold().write(u32::from(threshold));
+		}
+	}
+
+	pub fn claim_interrupt(&mut self) -> Option<NonZeroU16> {
+		trace!("Claiming interrupt from APLIC");
+		if let Some(msi_controller) = msi_controller() {
+			msi_controller.claim_interrupt()
+		} else {
+			let hart_idc = unsafe {
+				self.interrupt_delivery_control
+					.as_mut_ptr()
+					.map(|control| control.cast().offset(Aplic::get_hart_index() as isize))
+			};
+			let claimi = hart_idc.claimi().read();
+			NonZeroU16::new(claimi.identity())
+		}
+	}
+
+	pub fn complete_interrupt(&mut self, _irq_number: u16) {
+		if let Some(msi_controller) = msi_controller() {
+			msi_controller.complete_interrupt(NonZeroU16::new(_irq_number).unwrap());
+		} else {
+			// reading claimi register automatically completes the interrupt
+		}
+	}
+}
+
+pub fn init_aplic(addr: PhysAddr, size: usize, msi_delivery: bool) {
+	assert!(
+		addr.is_aligned_to(BasePageSize::SIZE),
+		"Aplic control region is not page aligned"
+	);
+	assert!(
+		size == 32 * 1024,
+		"Expected 32 KiB control region for APLIC in direct delivery mode"
+	);
+
+	let layout = PageLayout::from_size(size).unwrap();
+	let page_range = PageAlloc::allocate(layout).unwrap();
+	let control_region_addr = VirtAddr::from(page_range.start());
+
+	let mut flags = PageTableEntryFlags::empty();
+	flags.device().normal().writable().execute_disable();
+	paging::map::<BasePageSize>(
+		control_region_addr,
+		addr,
+		size / usize::try_from(BasePageSize::SIZE).unwrap(),
+		flags,
+	);
+
+	let control_region = unsafe {
+		VolatileRef::new(
+			NonNull::new(control_region_addr.as_mut_ptr::<AplicControlRegion>()).unwrap(),
+		)
+	};
+
+	let interrupt_delivery_control = unsafe {
+		VolatileRef::new(
+			NonNull::new(
+				VirtAddr::from(control_region_addr.as_u64() + 0x4000)
+					.as_mut_ptr::<InterruptDeliveryControlArray>(),
+			)
+			.unwrap(),
+		)
+	};
+	let mut aplic = Aplic::new(control_region, interrupt_delivery_control);
+	aplic.init(msi_delivery);
+
+	*EXTERNAL_INTERRUPT_CONTROLLER.lock() = Some(ExternalInterruptController::Aplic(aplic));
+}
